Share single Spring context across the master integration tests
Move the @SpringBootTest / @Import / inner @TestConfiguration declarations
of CommitIdPluginGitInformationProviderTest,
CommitIdPluginShortBuildInfoProviderTest and
DefaultBasicDiscoveryMetadataAssemblerTest into a new
AbstractMasterIntegrationTest base class, so the three tests produce an
identical merged context configuration and reuse one cached
ApplicationContext. AxelixMetadataEndpointTest is intentionally left out
of the hierarchy.
