+++ sbs/axelix-spring-boot-2-starter/src/test/java/com/axelixlabs/axelix/sbs/spring/core/master/AbstractMasterIntegrationTest.java
@@ -1,0 +1,79 @@
+/*
+ * Copyright (C) 2025-2026 Axelix Labs
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3 of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+ */
+package com.axelixlabs.axelix.sbs.spring.core.master;
+
+import java.lang.management.ManagementFactory;
+
+import org.springframework.boot.autoconfigure.info.ProjectInfoAutoConfiguration;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.context.TestConfiguration;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Import;
+
+import com.axelixlabs.axelix.common.api.registration.BasicDiscoveryMetadata;
+import com.axelixlabs.axelix.common.domain.version.AxelixVersionDiscoverer;
+
+/**
+ * Base class for the master-related integration tests.
+ *
+ * <p>Every master test extends this class and declares no context-affecting annotations of its own
+ * (no {@code @SpringBootTest}, {@code @TestPropertySource}, {@code @Import} or nested
+ * {@code @TestConfiguration} classes). As a result, all the master tests produce an identical
+ * merged context configuration and therefore share a single cached Spring application context,
+ * which is only started once for the whole test run.
+ *
+ * <p>{@link AxelixMetadataEndpointTest} is intentionally not part of this hierarchy.
+ *
+ * <p>The annotations below are the union of the configuration previously declared by the
+ * individual master tests.
+ */
+@SpringBootTest
+@Import({
+    CommitIdPluginGitInformationProvider.class,
+    CommitIdPluginShortBuildInfoProvider.class,
+    DefaultServiceMetadataAssembler.class,
+    ProjectInfoAutoConfiguration.class,
+    AbstractMasterIntegrationTest.MasterTestConfiguration.class
+})
+public abstract class AbstractMasterIntegrationTest {
+
+    @TestConfiguration
+    static class MasterTestConfiguration {
+
+        @Bean
+        VMFeaturesProvider vmFeaturesProvider() {
+            return new OptionsParsingVMFeaturesProvider(
+                    ManagementFactory.getRuntimeMXBean().getInputArguments());
+        }
+
+        @Bean
+        HealthDetectionFunction healthDetectionFunction() {
+            return () -> BasicDiscoveryMetadata.HealthStatus.UP;
+        }
+
+        @Bean
+        AxelixVersionDiscoverer axelixVersionDiscoverer() {
+            return () -> "1.1.3";
+        }
+
+        @Bean
+        LibraryInformationProvider libraryInformationProvider() {
+            return new DefaultLibraryInformationProvider();
+        }
+    }
+}
