+++ sbs/axelix-spring-boot-2-starter/src/test/java/com/axelixlabs/axelix/sbs/spring/core/master/CommitIdPluginGitInformationProviderTest.java
@@ -20,9 +20,6 @@ package com.axelixlabs.axelix.sbs.spring.core.master;
 import org.junit.jupiter.api.Test;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.autoconfigure.info.ProjectInfoAutoConfiguration;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.context.annotation.Import;
 
 import com.axelixlabs.axelix.common.api.registration.GitInfo;
 
@@ -33,9 +30,7 @@ import static org.assertj.core.api.Assertions.assertThat;
  *
  * @author Mikhail Polivakha
  */
-@SpringBootTest
-@Import({CommitIdPluginGitInformationProvider.class, ProjectInfoAutoConfiguration.class})
-class CommitIdPluginGitInformationProviderTest {
+class CommitIdPluginGitInformationProviderTest extends AbstractMasterIntegrationTest {
 
     @Autowired
     private CommitIdPluginGitInformationProvider subject;
